+++ Makefile
@@ -1,0 +1,2 @@
+run-ci:
+	black src/
