+++ README.md
@@ -1,0 +1,9 @@
+# Installation
+```bash
+git clone git@github.com:explodinggradients/synthetic-qa-paper.git
+cd synthetic-qa-paper
+pip install -e .
+```
+
+## Usage
+See example [notebook](notebooks/experiment.ipynb) for usage.
