+++ src/ragbench/__init__.py
@@ -1,0 +1,9 @@
+from ragbench.rag_utils import RAGEval, RAGTools, RAGPipeline
+from ragbench.customized_faiss import FAISS
+
+__all__ = [
+    "RAGEval",
+    "RAGTools",
+    "RAGPipeline",
+    "FAISS",
+]
